✨ Renamed Homepage to LandingPage as it's causing lint problems
